update after refactor + extract to file
